add date on which 75% of rewards begin to unlock
This was requested multiple times in the LuaSwap telegram channel by myself, Faruk, and Aadi

Please let me know if there is another format that my pull requests should take - I'm happy to conform to code contributing guidelines!  Also LMK if the date is correct.. I did it based on counting the weeks, but I could be wrong. Thanks!